flask/requirements.txt
requests
beautifulsoup4
pdfminer.six
python-docx
faiss-cpu
numpy
torch
transformers
flask
